@@ -22,28 +22,33 @@
   ],
   "scripts": {
     "bootstrap": "npm i && lerna bootstrap",
-    "serve": "node server",
-    "serve:ssr": "cross-env SSR=1 node --experimental-specifier-resolution=node server",
+    "serve": "npm run _node -- server",
+    "serve:ssr": "cross-env SSR=1 npm run _node -- server",
     "dev": "run-p serve watch:*",
     "dev:ssr": "run-p serve:ssr watch:*",
-    "lint": "eslint .",
-    "fixlint": "eslint . --fix",
-    "test": "jest --runInBand",
-    "test:e2e": "jest --selectProjects e2e",
-    "test:integration": "jest --selectProjects integration",
-    "test:unit": "jest --selectProjects unit",
-    "css": "node build/css",
+    "lint": "npm run _eslint -- .",
+    "fixlint": "npm run _eslint -- . --fix",
+    "test": "npm run _jest -- --runInBand",
+    "test:e2e": "npm run _jest -- --selectProjects e2e",
+    "test:integration": "npm run _jest -- --selectProjects integration",
+    "test:unit": "npm run _jest -- --selectProjects unit",
+    "css": "npm run _node -- build/css.js",
     "watch:css": "npm run css -- -o themes -w",
-    "watch:js": "node build/build.js",
-    "build:css:min": "mkdirp lib/themes && npm run css -- -o lib/themes && node build/mincss.js",
+    "watch:js": "npm run _node -- build/build.js",
+    "build:css:min": "mkdirp lib/themes && npm run css -- -o lib/themes && npm run _node -- build/mincss.js",
     "build:css": "mkdirp themes && npm run css -- -o themes",
-    "build:js": "cross-env NODE_ENV=production node build/build.js",
-    "build:cover": "node build/cover.js",
+    "build:js": "cross-env NODE_ENV=production npm run _node -- build/build.js",
+    "build:cover": "npm run _node -- build/cover.js",
     "build": "rimraf lib themes && run-s build:js build:css build:css:min build:cover",
     "prepare": "npm run build",
     "pub:next": "cross-env RELEASE_TAG=next sh build/release.sh",
     "pub": "sh build/release.sh",
-    "postinstall": "opencollective-postinstall"
+    "postinstall": "opencollective-postinstall",
+    "_node": "node --experimental-specifier-resolution=node",
+    "_eslint": "DISABLE_V8_COMPILE_CACHE=1 node -- ./node_modules/.bin/eslint",
+    "// ^ _eslint comment": "The DISABLE_V8_COMPILE_CACHE=1 var enables ESM import() in .eslintrc.cjs. See https://github.com/eslint/eslint/issues/13684",
+    "_jest": "npm run _node -- --experimental-vm-modules ./node_modules/.bin/jest",
+    "// ^ _jest comment": "--experimental-vm-modules is needed in order to enable full native ESM support in Jest"
   },
   "husky-OFF": {
     "hooks": {
@@ -70,7 +75,7 @@
     "@babel/preset-env": "^7.11.5",
     "autoprefixer-stylus": "^1.0.0",
     "axios": "^0.20.0",
-    "babel-eslint": "^10.0.3",
+    "babel-eslint": "10.0.3",
     "babel-jest": "^26.3.0",
     "browser-sync": "^2.26.12",
     "chokidar": "^3.4.2",
@@ -79,22 +84,25 @@
     "copy-dir": "^1.2.0",
     "cross-env": "^6.0.3",
     "cssnano": "^4.1.10",
-    "eslint": "^5.16.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jest": "^24.0.2",
-    "eslint-plugin-jest-playwright": "^0.2.1",
-    "eslint-plugin-prettier": "^3.1.2",
+    "deasync": "^0.1.20",
+    "es-main": "^1.0.2",
+    "eslint": "7.12.1",
+    "eslint-plugin-import": "2.20.1",
+    "eslint-plugin-jest": "24.0.2",
+    "eslint-plugin-jest-playwright": "0.2.1",
+    "eslint-plugin-prettier": "3.1.2",
     "husky": "^3.1.0",
-    "jest": "^26.4.2",
-    "jest-image-snapshot": "^4.2.0",
-    "jest-playwright-preset": "^1.3.1",
+    "jest": "26.4.2",
+    "jest-image-snapshot": "4.2.0",
+    "jest-playwright-preset": "1.3.1",
     "lerna": "^3.22.1",
     "lint-staged": "^10.4.0",
     "live-server": "^1.2.1",
     "mkdirp": "^0.5.1",
     "npm-run-all": "^4.1.5",
     "playwright": "^1.4.1",
-    "prettier": "^1.19.1",
+    "playwright-firefox": "^1.5.2",
+    "prettier": "2.1.2",
     "rimraf": "^3.0.0",
     "rollup": "^1.23.1",
     "rollup-plugin-async": "^1.2.0",
