@@ -1,10 +1,8 @@
-// Modules, constants, and variables
-// -----------------------------------------------------------------------------
-const docsifyInit = require('../helpers/docsify-init');
+import docsifyInit from '../helpers/docsify-init';
 
 // Suite
 // -----------------------------------------------------------------------------
-describe(`Example Tests`, function() {
+describe(`Example Tests`, function () {
   // Tests
   // ---------------------------------------------------------------------------
   test('dom manipulation', async () => {
