@@ -1,8 +1,8 @@
-const docsifyInit = require('../helpers/docsify-init');
+import docsifyInit from '../helpers/docsify-init';
 
 // Suite
 // -----------------------------------------------------------------------------
-describe('Search Plugin Tests', function() {
+describe('Search Plugin Tests', function () {
   // Tests
   // ---------------------------------------------------------------------------
   test('search readme', async () => {
