@@ -1,19 +1,35 @@
 // @ts-check
-
-// const { initJSDOM } = require('../_helper');
-
-// const port = 9754;
-// const docsifySite = 'http://127.0.0.1:' + port;
-
-// initJSDOM();
-
+import fs from 'fs';
 import {
   Renderer,
   getServerHTMLTemplate,
 } from '../../packages/docsify-server-renderer/index';
 
-describe('pacakges/docsify-server-render', function() {
-  it('renders content', async function() {
+describe('pacakges/docsify-server-render Renderer', function () {
+  it('renders content', async function () {
+    const renderer = new Renderer({
+      template: getServerHTMLTemplate(),
+      config: {
+        name: 'docsify',
+        repo: 'docsifyjs/docsify',
+        // basePath: 'https://docsify.js.org/',
+        loadNavbar: true,
+        loadSidebar: true,
+        subMaxLevel: 3,
+        auto2top: true,
+      },
+    });
+
+    expect(renderer).toBeInstanceOf(Renderer);
+
+    const result = await renderer.renderToString('/docs/ssr.md');
+
+    expect(typeof result).toBe('string');
+    expect(result).toContain('<h1 id="server-side-rendering">');
+    expect(result).toContain('<span>Server-Side Rendering</span></a></h1>');
+  });
+
+  it('renders aliases the same', async () => {
     const renderer = new Renderer({
       template: getServerHTMLTemplate(),
       config: {
@@ -27,14 +43,24 @@ describe('pacakges/docsify-server-render', function() {
         alias: {
           '/de-de/changelog': '/changelog',
           '/zh-cn/changelog': '/changelog',
-          '/changelog':
-            'https://raw.githubusercontent.com/docsifyjs/docsify/master/CHANGELOG',
+          '/changelog': '/docs/changelog.md',
         },
       },
     });
 
-    await renderer.renderToString('/changelog');
+    const [one, two, three] = await Promise.all([
+      await renderer.renderToString('/de-de/changelog'),
+      await renderer.renderToString('/zh-cn/changelog'),
+      await renderer.renderToString('/changelog'),
+    ]);
 
-    expect(renderer).toBeInstanceOf(Renderer);
+    await Promise.all([
+      fs.promises.writeFile('tmp1.html', one),
+      fs.promises.writeFile('tmp2.html', two),
+      fs.promises.writeFile('tmp3.html', three),
+    ]);
+
+    expect(one).toEqual(two);
+    expect(two).toEqual(three);
   });
 });
