@@ -60,7 +60,7 @@ export function fetchMixin(proto) {
     return path404;
   };
 
-  proto._loadSideAndNav = function(path, qs, loadSidebar, cb) {
+  proto._loadSideAndNav = function (path, qs, loadSidebar, cb) {
     return () => {
       if (!loadSidebar) {
         return cb();
@@ -76,7 +76,7 @@ export function fetchMixin(proto) {
     };
   };
 
-  proto._fetch = function(cb = noop) {
+  proto._fetch = function (cb = noop) {
     const { path, query } = this.route;
     const qs = stringifyQuery(query, ['id']);
     const { loadNavbar, requestHeaders, loadSidebar } = this.config;
@@ -114,7 +114,7 @@ export function fetchMixin(proto) {
       );
   };
 
-  proto._fetchCover = function() {
+  proto._fetchCover = function () {
     const { coverpage, requestHeaders } = this.config;
     const query = this.route.query;
     const root = getParentPath(this.route.path);
@@ -150,7 +150,7 @@ export function fetchMixin(proto) {
     }
   };
 
-  proto.$fetch = function(
+  proto.$fetch = function (
     cb = noop,
     $resetEvents = this.$resetEvents.bind(this)
   ) {
@@ -171,7 +171,7 @@ export function fetchMixin(proto) {
     }
   };
 
-  proto._fetchFallbackPage = function(path, qs, cb = noop) {
+  proto._fetchFallbackPage = function (path, qs, cb = noop) {
     const { requestHeaders, fallbackLanguages, loadSidebar } = this.config;
 
     if (!fallbackLanguages) {
@@ -210,7 +210,7 @@ export function fetchMixin(proto) {
    * @returns {Boolean} True if the requested page is not found
    * @private
    */
-  proto._fetch404 = function(path, qs, cb = noop) {
+  proto._fetch404 = function (path, qs, cb = noop) {
     const { loadSidebar, requestHeaders, notFoundPage } = this.config;
 
     const fnLoadSideAndNav = this._loadSideAndNav(path, qs, loadSidebar, cb);
