@@ -10,7 +10,7 @@ import { get } from './fetch/ajax';
 // major release. We'll tell people to get everything from the DOCSIFY global
 // when using the global build, but we'll highly recommend for them to import
 // from the ESM build (f.e. lib/docsify.esm.js and lib/docsify.min.esm.js).
-export default function() {
+export default function () {
   window.Docsify = {
     util,
     dom,
