@@ -3,7 +3,7 @@ import Prism from 'prismjs';
 import 'prismjs/components/prism-markup-templating';
 
 export const highlightCodeCompiler = ({ renderer }) =>
-  (renderer.code = function(code, lang = 'markup') {
+  (renderer.code = function (code, lang = 'markup') {
     const langOrMarkup = Prism.languages[lang] || Prism.languages.markup;
     const text = Prism.highlight(
       code.replace(/@DOCSIFY_QM@/g, '`'),
