@@ -13,7 +13,7 @@ const CONFIG = {
   pathNamespaces: undefined,
 };
 
-const install = function(hook, vm) {
+const install = function (hook, vm) {
   const { util } = Docsify;
   const opts = vm.config.search || CONFIG;
 
